backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.12.0
aiofiles==23.2.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# OAuth
authlib==1.2.1
requests==2.31.0

# Validation
pydantic[email]==2.5.0
